moved public inner classes and interfaces in OrientDB Test Commons to appropriate places
