+++ core/src/test/java/com/orientechnologies/orient/core/index/hashindex/local/cache/ConcurrentLRUListConcurrentTest.java
@@ -12,6 +12,7 @@ import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
 import com.orientechnologies.orient.test.ConcurrentTestHelper;
+import com.orientechnologies.orient.test.TestFactory;
 
 /**
  * Concurrent test for {@link ConcurrentLRUList}.
@@ -112,7 +113,7 @@ public class ConcurrentLRUListConcurrentTest {
     c = c1;
   }
 
-  private class AdderFactory implements ConcurrentTestHelper.TestFactory<Integer> {
+  private class AdderFactory implements TestFactory<Integer> {
     private int j = 0;
 
     @Override
@@ -131,7 +132,7 @@ public class ConcurrentLRUListConcurrentTest {
     }
   }
 
-  private class RemoveLRUFactory implements ConcurrentTestHelper.TestFactory<Integer> {
+  private class RemoveLRUFactory implements TestFactory<Integer> {
     @Override
     public Callable<Integer> createWorker() {
       return new Callable<Integer>() {
@@ -152,7 +153,7 @@ public class ConcurrentLRUListConcurrentTest {
     }
   }
 
-  private class RandomAdderFactory implements ConcurrentTestHelper.TestFactory<Integer> {
+  private class RandomAdderFactory implements TestFactory<Integer> {
 
     @Override
     public Callable<Integer> createWorker() {
@@ -171,7 +172,7 @@ public class ConcurrentLRUListConcurrentTest {
     }
   }
 
-  private class AddSameFactory implements ConcurrentTestHelper.TestFactory<Integer> {
+  private class AddSameFactory implements TestFactory<Integer> {
 
     @Override
     public Callable<Integer> createWorker() {
@@ -190,7 +191,7 @@ public class ConcurrentLRUListConcurrentTest {
     }
   }
 
-  private class RandomRemoveFactory implements ConcurrentTestHelper.TestFactory<Integer> {
+  private class RandomRemoveFactory implements TestFactory<Integer> {
     @Override
     public Callable<Integer> createWorker() {
       return new Callable<Integer>() {
