+++ test-commons/src/main/java/com/orientechnologies/orient/test/TestFactory.java
@@ -1,0 +1,30 @@
+/*
+ * Copyright 2015 Orient Technologies.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.orientechnologies.orient.test;
+
+import java.util.concurrent.Callable;
+
+/**
+ *
+ * @author richter
+ * @param <T> the type of the {@link Callable} to be created with 
+ * {@link #createWorker() }
+ */
+public interface TestFactory<T> {
+
+  Callable<T> createWorker();
+    
+}
